Fix Sonar warnings.
