+++ refinedstorage2-network-api/src/test/java/com/refinedmods/refinedstorage2/api/network/node/diskdrive/DiskDriveNetworkNodeTest.java
@@ -24,20 +24,14 @@ import java.util.Set;
 import java.util.UUID;
 
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EnumSource;
 import org.junit.jupiter.params.provider.ValueSource;
-import org.mockito.verification.VerificationMode;
 
 import static com.refinedmods.refinedstorage2.network.test.nodefactory.AbstractNetworkNodeFactory.PROPERTY_ENERGY_USAGE;
 import static com.refinedmods.refinedstorage2.network.test.nodefactory.DiskDriveNetworkNodeFactory.PROPERTY_ENERGY_USAGE_PER_DISK;
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.never;
-import static org.mockito.Mockito.times;
-import static org.mockito.Mockito.verify;
 
 @NetworkTest
 @SetupNetwork
@@ -52,14 +46,11 @@ class DiskDriveNetworkNodeTest {
     DiskDriveNetworkNode sut;
 
     FakeStorageProviderRepository storageProviderRepository;
-    DiskDriveListener diskDriveListener;
 
     @BeforeEach
     void setUp() {
-        diskDriveListener = mock(DiskDriveListener.class);
         storageProviderRepository = new FakeStorageProviderRepository();
         sut.setDiskProvider(storageProviderRepository);
-        sut.setListener(diskDriveListener);
     }
 
     @Test
@@ -634,81 +625,6 @@ class DiskDriveNetworkNodeTest {
     }
 
     @Test
-    void shouldNotCallDiskStateChangeListenerWhenUnnecessaryOnExtracting(
-        @InjectNetworkStorageChannel final StorageChannel<String> networkStorage
-    ) {
-        // Arrange
-        final Storage<String> storage = new LimitedStorageImpl<>(100);
-        storage.insert("A", 76, Action.EXECUTE, EmptyActor.INSTANCE);
-        storageProviderRepository.setInSlot(1, storage);
-        initializeDiskDriveAndActivate();
-
-        // Act
-        networkStorage.extract("A", 1, Action.EXECUTE, EmptyActor.INSTANCE);
-
-        // Assert
-        verify(diskDriveListener, never()).onDiskChanged();
-    }
-
-    @Test
-    void shouldNotCallDiskStateChangeListenerWhenUnnecessaryOnInserting(
-        @InjectNetworkStorageChannel final StorageChannel<String> networkStorage
-    ) {
-        // Arrange
-        final Storage<String> storage = new LimitedStorageImpl<>(100);
-        storageProviderRepository.setInSlot(1, storage);
-        initializeDiskDriveAndActivate();
-
-        // Act
-        networkStorage.insert("A", 74, Action.EXECUTE, EmptyActor.INSTANCE);
-
-        // Assert
-        verify(diskDriveListener, never()).onDiskChanged();
-    }
-
-    @ParameterizedTest
-    @EnumSource(Action.class)
-    void shouldCallDiskStateChangeListenerWhenExtracting(
-        final Action action,
-        @InjectNetworkStorageChannel final StorageChannel<String> networkStorage
-    ) {
-        // Arrange
-        final Storage<String> storage = new LimitedStorageImpl<>(100);
-        storage.insert("A", 75, Action.EXECUTE, EmptyActor.INSTANCE);
-        storageProviderRepository.setInSlot(1, storage);
-        initializeDiskDriveAndActivate();
-
-        // Act
-        networkStorage.extract("A", 1, action, EmptyActor.INSTANCE);
-        networkStorage.extract("A", 1, action, EmptyActor.INSTANCE);
-
-        // Assert
-        final VerificationMode expectedTimes = action == Action.EXECUTE ? times(1) : never();
-        verify(diskDriveListener, expectedTimes).onDiskChanged();
-    }
-
-    @ParameterizedTest
-    @EnumSource(Action.class)
-    void shouldCallDiskStateChangeListenerWhenInserting(
-        final Action action,
-        @InjectNetworkStorageChannel final StorageChannel<String> networkStorage
-    ) {
-        // Arrange
-        final Storage<String> storage = new LimitedStorageImpl<>(100);
-        storage.insert("A", 74, Action.EXECUTE, EmptyActor.INSTANCE);
-        storageProviderRepository.setInSlot(1, storage);
-        initializeDiskDriveAndActivate();
-
-        // Act
-        networkStorage.insert("A", 1, action, EmptyActor.INSTANCE);
-        networkStorage.insert("A", 1, action, EmptyActor.INSTANCE);
-
-        // Assert
-        final VerificationMode expectedTimes = action == Action.EXECUTE ? times(1) : never();
-        verify(diskDriveListener, expectedTimes).onDiskChanged();
-    }
-
-    @Test
     void shouldTrackChanges(@InjectNetworkStorageChannel final StorageChannel<String> networkStorage) {
         // Arrange
         final Storage<String> storage = new TrackedStorageImpl<>(new LimitedStorageImpl<>(100), () -> 0L);
@@ -726,48 +642,5 @@ class DiskDriveNetworkNodeTest {
     private void initializeDiskDriveAndActivate() {
         sut.initialize(storageProviderRepository);
         sut.setActive(true);
-    }
-
-    @Nested
-    class PriorityTest {
-        @AddNetworkNode
-        DiskDriveNetworkNode otherDiskDrive;
-
-        @ParameterizedTest
-        @ValueSource(booleans = {true, false})
-        void shouldRespectPriority(final boolean oneHasPriority,
-                                   @InjectNetworkStorageChannel final StorageChannel<String> networkStorage) {
-            // Arrange
-            final Storage<String> storage1 = new LimitedStorageImpl<>(100);
-            storageProviderRepository.setInSlot(1, storage1);
-            initializeDiskDriveAndActivate();
-
-            final Storage<String> storage2 = new LimitedStorageImpl<>(100);
-            final FakeStorageProviderRepository storageProviderManager2 = new FakeStorageProviderRepository();
-            storageProviderManager2.setInSlot(1, storage2);
-            otherDiskDrive.setDiskProvider(storageProviderManager2);
-            otherDiskDrive.initialize(storageProviderManager2);
-            otherDiskDrive.setActive(true);
-
-            if (oneHasPriority) {
-                sut.setPriority(5);
-                otherDiskDrive.setPriority(2);
-            } else {
-                sut.setPriority(2);
-                otherDiskDrive.setPriority(5);
-            }
-
-            // Act
-            networkStorage.insert("A", 1, Action.EXECUTE, EmptyActor.INSTANCE);
-
-            // Assert
-            if (oneHasPriority) {
-                assertThat(storage1.getAll()).isNotEmpty();
-                assertThat(storage2.getAll()).isEmpty();
-            } else {
-                assertThat(storage1.getAll()).isEmpty();
-                assertThat(storage2.getAll()).isNotEmpty();
-            }
-        }
     }
 }
